requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
dataclasses==0.6
Django==2.2.4
django-rest-knox==4.1.0
djangorestframework==3.10.2
djongo==1.2.33
pkg-resources==0.0.0
pycparser==2.19
pymongo==3.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.2.4
